ch592 pm: capture and clear the post latch with one amoswap.w in pm_loop_top
Review of PR #36 (CodeRabbit, Copilot): the loop top read dongle_pm_post
and then cleared it, so an ISR post landing between the two was erased
from the latch. The event itself stayed queued in TMOS and this
iteration's three passes dispatch it, so no sleep-admission hole opened
under the current scheduler bound (the same race existed in the loop-top
code this branch moved), but the latch is what the entry alarm and the
quiet-pass veto read.

The latch is now a word and the loop top swaps it with zero in one
amoswap.w (the A extension of this rv32imac build; the objdump shows the
single instruction): a post lands either before the swap and is captured,
or after it and survives for the next loop top. Not an interrupt mask:
taking CSR 0x800 on every main-loop iteration is the enumeration hazard the
USB_ServiceRemoteWake comment in usb_device.c records. dongle_pm_ran is
written by task context only and pm_entry_armed by this loop only, so
their plain clears race nothing.

All-off image unchanged (c48fab05... with BUILD_ID forced). The admission
logic is the one measured on the bench in the previous commit; only the
capture instruction changed.

+++ firmware/ch592/src/pm_ch592.h
@@ -15,9 +15,10 @@
 
 #if DONGLE_PM_IDLE
 /* Per-iteration "work happened" latches (why: pm_ch592.c header comment).
- * Plain byte stores from every context; cleared only by pm_loop_top at the
- * top of Main_Circulation. */
-extern volatile uint8_t dongle_pm_post;   /* set by every hal_event_post   */
+ * Plain stores from every context; cleared only by pm_loop_top at the top of
+ * Main_Circulation (the post latch with one amoswap, so no ISR post is lost
+ * between a read and a clear). */
+extern volatile uint32_t dongle_pm_post;  /* set by every hal_event_post; a word so the loop top can amoswap it */
 extern volatile uint8_t dongle_pm_ran;    /* set at RF_ProcessEvent entry  */
 /* The host LED state poll_usb_led_state (main.c) last relayed; the masked
  * admission re-check compares the live USB snapshot against it. */
